// src/infrastructure/parameters/types.ts
export type Parameters = {
  withdrawalTronWalletAlertLevel?: number;
  withdrawalTronUsdtWalletAlertLevel?: number;
  feesTronWalletAlertLevel?: number;
  withdrawalTronUsdcWalletAlertLevel?: number;
  withdrawalSolanaWalletAlertLevel?: number;
  withdrawalSolanaUsdtWalletAlertLevel?: number;
  withdrawalSolanaUsdcWalletAlertLevel?: number;
  feesSolanaWalletAlertLevel?: number;
  withdrawalEthWalletAlertLevel?: number;
  withdrawalEthUsdtWalletAlertLevel?: number;
  withdrawalEthUsdcWalletAlertLevel?: number;
};